Task C: Catalog Display - Iteration C1
Iteration C1: Creating the Catalog Listing

/app/controllers/store_controller.rb
@products = Product.order(:title)

/app/views/store/index.html.erb
write the view template. To do this, edit the file index.html.erb in
app/views/store.
Note the use of the sanitize() method for the description. This allows
us to safely add HTML stylings to make the descriptions more
interesting for our customers. Note that this decision opens a
potential security hole,1 but because product descriptions are created
by people who work for our company, we think that the risk is minimal.
We’ve also used the image_tag() helper method. This generates an HTML
<img> tag using its argument as the image source.