Task C: Catalog Display - Iteration C1
Iteration C1: Creating the Catalog Listing

/app/assets/stylesheets/store.css.scss

It is still pretty basic, and it seems to be missing something. The
customer happens to be walking by as we ponder this, and she points out
that she’d also like to see a decent-looking banner and sidebar on
public-facing pages.
At this point in the real world, we’d probably want to call in the
design folks