Task D: Cart Creation - Iteration D1
Iteration D1: Finding a Cart

Rails makes the current session look like a hash to the controller, so
we’ll store the ID of the cart in the session by indexing it with the
symbol :cart_id.

/app/controllers/concerns/current_cart.rb

The set_cart() method starts by getting the :cart_id from the session
object and then attempts to find a cart corresponding to this ID. If
such a cart record is not found (which will happen if the ID is nil or
invalid for any reason), then this method will proceed to create a new
Cart, store the ID of the created cart into the session, and then
return the new cart.

place the set_cart() method in a CurrentCart module and mark it as
private. This treatment allows us to share common code (even as little
as a single method!) between controllers and furthermore prevents Rails
from ever making it available as an action on the controller.