Task D: Cart Creation - Iteration D2
Iteration D2: Connecting Products to Carts

/app/models/line_item.rb
At the model level, there is no difference between a simple reference
and a “belongs to” relationship. Both are implemented using the
belongs_to() method.
There’s an easy way to remember where to put belongs_to declarations:
if a table has foreign keys, the corresponding model should have a
belongs_to for each.
li = LineItem.find(...)
puts "This line item is for #{li.product.title}"

/app/models/cart.rb
That has_many :line_items part of the directive is fairly
self-explanatory: a cart (potentially) has many associated line items.
These are linked to the cart because each line item contains a
reference to its cart’s ID. The dependent: :destroy part indicates that
the existence of line items is dependent on the existence of the cart.
If we destroy a cart, deleting it from the database, we’ll want Rails
also to destroy any line items that are associated with that cart.
Now that the Cart is declared to have many line items, we can reference
them (as a collection) from a cart object.
cart = Cart.find(...)
puts "This cart has #{cart.line_items.count} line items"