Task D: Cart Creation - Iteration D2
Iteration D2: Connecting Products to Carts

/app/models/product.rb
Here we declare that a product has many line items and define a hook
method named ensure_not_referenced_by_any_line_item(). A hook method is
a method that Rails calls automatically at a given point in an object’s
life. In this case, the method will be called before Rails attempts to
destroy a row in the database. If the hook method returns false, the
row will not be destroyed.