Task D: Cart Creation - Iteration D3
Iteration D3: Adding a Button

app/controllers/line_items_controller.rb
We use the params object to get the :product_id parameter from the
request. The params object is important inside Rails applications. It
holds all of the param- eters passed in a browser request. We store the
result in a local variable because there is no need to make this
available to the view.
We then pass that product we found into @cart.line_items.build. This
causes a new line item relationship to be built between the @cart
object and the product. You can build the relationship from either end,
and Rails will take care of establishing the connections on both sides.
We save the resulting line item into an instance variable named
@line_item.
once the line item is created, we want to redirect you to the cart
instead of back to the line item. Since the line item object knows how
to find the cart object, all we need to do is add .cart to the method
call.