Task E: A Smarter Cart
Iteration E1: Creating a Smarter Cart

Rails can tell from the name of the migration that you are adding one
or more columns to the line_items table and can pick up the names and
data types for each column from the last argument. The two patterns
that Rails matches on are add_XXX_to_TABLE and remove_XXX_from_TABLE,
where the value of XXX is ignored; what matters is the list of column
names and types that appear after the migration name.
The only thing Rails can’t tell is what a reasonable default is for
this column.

In many cases, a null value would do, but let’s make it the value 1 for
existing carts by modifying the migration before we apply it.
db/migrate/20150127083429_add_quantity_to_line_items.rb