Task E: A Smarter Cart
Iteration E1: Creating a Smarter Cart

/app/models/cart.rb
 we need a smart add_product() method in our Cart, one that checks
whether our list of items already includes the product we’re adding; if
it does, it bumps the quantity, and if it doesn’t, it builds a new
LineItem.
The find_by() method is a streamlined version of the where() method.
Instead of returning an array of results, it returns either an existing
LineItem or nil.