Task E: A Smarter Cart
Iteration E1: Creating a Smarter Cart

• We start by iterating over each cart.
• For each cart, we get a sum of the quantity fields for each of the
line items associated with this cart, grouped by product_id. The
resulting sums will be a list of ordered pairs of product_ids and
quantity.
• We iterate over these sums, extracting the product_id and quantity
from each.
• In cases where the quantity is greater than one, we will delete all
of the individual line items associated with this cart and this product
and replace them with a single line item with the correct quantity.