Task E: A Smarter Cart
Iteration E1: Creating a Smarter Cart

An important principle of migrations is that each step needs to be
reversible, so we implement a down() too. This method finds line items
with a quantity of greater than one; adds new line items for this cart
and product, each with a quantity of one; and finally deletes the line
item.