Task E: A Smarter Cart
Iteration E2: Handling Errors

A flash is a bucket (actually closer to a Hash) in which you can store
stuff as you process a request. The contents of the flash are available
to the next request in this session before being deleted automatically.
Typically the flash is used to collect error messages.

Armed with all this background about flash data, create an
invalid_cart() method to report on the problem.

/app/controllers/carts_controller.rb
The rescue_from clause intercepts the exception raised by Cart.find().
In the han- dler, we do the following:
• Use the Rails logger to record the error. Every controller has a
logger attribute. Here we use it to record a message at the error
logging level.
• Redirect to the catalog display using the redirect_to() method.

Started GET "/carts/wibble" for 127.0.0.1 at 2015-01-27 17:30:22 +0800
Processing by CartsController#show as HTML
  Parameters: {"id"=>"wibble"}
  [1m[35mCart Load (0.1ms)[0m  SELECT "carts".* FROM "carts" WHERE
"carts"."id" = ? LIMIT 1  [["id", "wibble"]]
>Attempt to access invalid cart wibble
Redirected to http://localhost:3000/
Completed 302 Found in 2ms (ActiveRecord: 0.1ms)