Task F: Add a Dash of Ajax - Iteration F1
Iteration F1: Moving the Cart

Ajax (which once stood for Asynchronous JavaScript and XML but now just
means “making browsers suck less”).

think of Rails partial templates (partials for short) as a kind of
method for views. A partial is simply a chunk of a view in its own
separate file. You can invoke (render) a partial from another template
or from a controller, and the partial will render itself and return the
results of that rendering. And, just as with methods, you can pass
parameters to a partial, so the same partial can render different
results.

The render() method will iterate over any collection that is passed to
it. The partial template itself is simply another template file (by
default in the same directory as the object being rendered and with the
name of the table as the name). However, to keep the names of partials
distinct from regular templates, Rails automatically prepends an
underscore to the partial name when looking for the file. That means we
need to name our partial _line_item.html.erb and place it in the
app/views/line_items directory.