Task F: Add a Dash of Ajax - Iteration F2
Iteration F2: Creating an Ajax-Based Cart

JavaScript Object Notation [JSON]

On the index page, we’re using button_to() to create the link to the
create action. We want to change this to send an Ajax request instead.
To do this, we simply add a remote: true parameter to the call.
/app/views/store/index.html.erb
 arranged for the browser to send an Ajax request to our applica- tion.