Task F: Add a Dash of Ajax - Iteration F2
Iteration F2: Creating an Ajax-Based Cart

have the application return a response. The plan is to create the
updated HTML fragment that represents the cart and to have the browser
stick that HTML into the browser’s internal representation of the
structure and content of the document being displayed, namely, the
Document Object Model (DOM). By manipulating the DOM, we cause the
display to change in front of the user’s eyes.

he first change is to stop the create action from redirecting to the
index display if the request is for JavaScript. We do this by adding a
call to respond_to() telling it that we want to respond with a format
of .js.

app/controllers/line_items_controller.rb