Task F: Add a Dash of Ajax - Iteration F2
Iteration F2: Creating an Ajax-Based Cart

Rails supports templates that generate JavaScript—the JS stands for
Java- Script. A .js.erb template is a way of getting JavaScript on the
browser to do what you want, all by writing server-side Ruby code.

/app/views/line_items/create.js.erb
This simple template tells the browser to replace the content of the
element whose id="cart" with that HTML.

For simplicity and conciseness, the jQuery library is aliased to $, and
most usages of jQuery start there.
The first call—$('#cart')—tells jQuery to find the HTML element that
has an id of cart. The html() method1 is then called with a first
argument of the desired replacement for the contents of this element.
This content is formed by calling the render() method on the @cart
object. The output of this method is processed by a escape_javascript()
helper method that converts this Ruby string into a format acceptable
as input to JavaScript.
Note that this script is executed in the browser. The only parts
executed on the server are the portions within the <%= and %>
delimiters.